factors out the backup register functions from the bootloader into the HAL so we can use them from system firmware
