+++ bootloader/src/core/bkpreg_hal.c
@@ -1,0 +1,3 @@
+
+
+#include "../src/core/bkpreg_hal.c"
