+++ bootloader/src/stm32f2xx/bkpreg_hal.c
@@ -1,0 +1,4 @@
+
+// pull in the sources from the HAL. It's a bit of a hack, but is simpler than trying to link the full
+// hal library.
+#include "../src/stm32f2xx/bkpreg_hal.c"
